requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==1.10.11
sentence-transformers==2.2.2
faiss-cpu==1.12.0
numpy>=1.26.4
requests==2.31.0
